Uat 1.0.126

+++ index.js
@@ -10,19 +10,32 @@ import {Provider} from 'react-redux';
 import {PersistGate} from 'redux-persist/integration/react';
 import ProviderFormattingContext from './src/context/FormattingContext';
 import {
+  serviceGetCustomCategories,
   thunkGetCustomCategories,
   thunkGetTags,
 } from './src/redux/category/categoryReducer';
 import {fetchGetAllSettings} from './src/redux/other/otherReducer';
-import {thunkGetSellPoints} from './src/redux/sellPoints/sellPointsReducer';
+import {
+  thunkGetExpressSellPoints,
+  thunkGetSellPoints,
+} from './src/redux/sellPoints/sellPointsReducer';
 import I18n from 'react-native-i18n';
 import en from './src/assets/translations/en';
 import uk from './src/assets/translations/uk';
+import ru from './src/assets/translations/ru';
 import {actionsUser, refreshUser} from './src/redux/user/userReducer';
 import service from './src/services/service';
 import {actionsCart} from './src/redux/cart/cartReducer';
 import {DEFAULT_NAME_SETTING} from './src/constants/constantsId';
 import {Host} from 'react-native-portalize';
+import messaging from '@react-native-firebase/messaging';
+import {
+  switchHandlerMessaging,
+  TypeHandlerMessaging,
+} from './src/useHooks/useHandlerMessaging';
+import {isIOS} from './src/utils/isPlatform';
+import {thunkGetTypes} from './src/redux/types/typeReducer';
+import AsyncStorage from '@react-native-community/async-storage';
 
 I18n.defaultLocale = 'uk';
 I18n.fallbacks = true;
@@ -30,30 +43,51 @@ I18n.fallbacks = true;
 I18n.translations = {
   uk,
   en,
+  ru,
 };
+
+messaging().setBackgroundMessageHandler(async (remoteMessage) => {
+  console.log(
+    '------------------>    ' +
+      (isIOS ? 'IOS' : 'ANDROID') +
+      ': Message handled in the background!',
+    remoteMessage,
+  );
+  switchHandlerMessaging({
+    type: TypeHandlerMessaging.background,
+    payload: remoteMessage,
+  });
+});
 
 const store = configureStore();
 store.store.dispatch(thunkGetCustomCategories);
+store.store.dispatch(serviceGetCustomCategories);
 store.store.dispatch(thunkGetTags);
 store.store.dispatch(fetchGetAllSettings);
 store.store.dispatch(thunkGetSellPoints);
+store.store.dispatch(thunkGetTypes);
 
-const handleAppStateChange = (nextAppState) => {
-  if (nextAppState === 'background' || nextAppState === 'inactive') {
-    const items = store.store.getState().cart.data;
-    const isAuth = store.store.getState().user.isAuth;
-    const id = store.store.getState().cart.idSellPoint
-      ? store.store.getState().cart.idSellPoint
-      : store.store.getState().other.settings[DEFAULT_NAME_SETTING]
-          .default_price_sell_point;
+const handleAppStateChange = async (nextAppState) => {
+  try {
+    if (nextAppState === 'background' || nextAppState === 'inactive') {
+      const root = store.store.getState();
+      const isAuth = root.user.isAuth;
+      if (!isAuth) return;
 
-    if (!isAuth) return;
+      const items = root.cart.data;
+      const idDeliveryType = root.order.deliveryType?.id;
+      const id = root.cart.idSellPoint
+        ? root.cart.idSellPoint
+        : root.other.settings[DEFAULT_NAME_SETTING].default_price_sell_point;
 
-    if (items.length > 0) {
-      service.saveCart(items, id);
-    } else {
-      service.deleteCart();
+      if (items.length > 0 && idDeliveryType) {
+        service.saveCart(items, id, idDeliveryType);
+      } else {
+        service.deleteCart();
+      }
     }
+  } catch (e) {
+    console.log({e});
   }
 };
 
@@ -63,8 +97,11 @@ export const getToken = () => {
   return store.store.getState().user.token;
 };
 
-export const getLocale = () => {
-  return store.store.getState().other.locale;
+export const getLocale = async () => {
+  return (
+    (await AsyncStorage.getItem('locale')) ||
+    store.store.getState().other.locale
+  );
 };
 
 export const logOut = () => {
